+++ internal/error_decoder.go
@@ -8,7 +8,7 @@ import (
 	"io"
 	"net/http"
 
-	"github.com/KardFinancial/kard-go-sdk/v20/core"
+	"github.com/KardFinancial/kard-go-sdk/v21/core"
 )
 
 // ErrorCodes maps HTTP status codes to error constructors.