+++ notifications/subscriptions/client.go
@@ -6,11 +6,11 @@ import (
 	context "context"
 	os "os"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v20"
-	core "github.com/KardFinancial/kard-go-sdk/v20/core"
-	internal "github.com/KardFinancial/kard-go-sdk/v20/internal"
-	notifications "github.com/KardFinancial/kard-go-sdk/v20/notifications"
-	option "github.com/KardFinancial/kard-go-sdk/v20/option"
+	kard "github.com/KardFinancial/kard-go-sdk/v21"
+	core "github.com/KardFinancial/kard-go-sdk/v21/core"
+	internal "github.com/KardFinancial/kard-go-sdk/v21/internal"
+	notifications "github.com/KardFinancial/kard-go-sdk/v21/notifications"
+	option "github.com/KardFinancial/kard-go-sdk/v21/option"
 )
 
 type Client struct {