+++ notifications/subscriptions/notifications_subscriptions_test/notifications_subscriptions_test.go
@@ -10,10 +10,10 @@ import (
 	os "os"
 	testing "testing"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v20"
-	client "github.com/KardFinancial/kard-go-sdk/v20/client"
-	notifications "github.com/KardFinancial/kard-go-sdk/v20/notifications"
-	option "github.com/KardFinancial/kard-go-sdk/v20/option"
+	kard "github.com/KardFinancial/kard-go-sdk/v21"
+	client "github.com/KardFinancial/kard-go-sdk/v21/client"
+	notifications "github.com/KardFinancial/kard-go-sdk/v21/notifications"
+	option "github.com/KardFinancial/kard-go-sdk/v21/option"
 	require "github.com/stretchr/testify/require"
 )