+++ organizations/children/organizations_children_test/organizations_children_test.go
@@ -10,9 +10,9 @@ import (
 	os "os"
 	testing "testing"
 
-	client "github.com/KardFinancial/kard-go-sdk/v20/client"
-	option "github.com/KardFinancial/kard-go-sdk/v20/option"
-	organizations "github.com/KardFinancial/kard-go-sdk/v20/organizations"
+	client "github.com/KardFinancial/kard-go-sdk/v21/client"
+	option "github.com/KardFinancial/kard-go-sdk/v21/option"
+	organizations "github.com/KardFinancial/kard-go-sdk/v21/organizations"
 	require "github.com/stretchr/testify/require"
 )