+++ organizations/contentstrategies/organizations_content_strategies_test/organizations_content_strategies_test.go
@@ -10,10 +10,10 @@ import (
 	os "os"
 	testing "testing"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v20"
-	client "github.com/KardFinancial/kard-go-sdk/v20/client"
-	option "github.com/KardFinancial/kard-go-sdk/v20/option"
-	organizations "github.com/KardFinancial/kard-go-sdk/v20/organizations"
+	kard "github.com/KardFinancial/kard-go-sdk/v21"
+	client "github.com/KardFinancial/kard-go-sdk/v21/client"
+	option "github.com/KardFinancial/kard-go-sdk/v21/option"
+	organizations "github.com/KardFinancial/kard-go-sdk/v21/organizations"
 	require "github.com/stretchr/testify/require"
 )